Implementing withdraw, some refactoring

+++ src/hu/csercsak_albert/banking_system/main/User.java
@@ -1,4 +1,4 @@
 package hu.csercsak_albert.banking_system.main;
 
-public record User(long id, String username, String firstName, String lastName, String email, int accountNumber) {
+public record User(int id, String username, String firstName, String lastName, String email, int accountNumber) {
 }
